create decision tree algorithm for attempt marking move actual code into the algorithm update psycopg2
