+++ polls/models/algorithm.py
@@ -2,10 +2,14 @@ import json
 import subprocess
 from django.db import models
 from .utils import class_import
+from ..script.sage_client import SageCell
+
 
 def algorithm_base_generate(atype, **kwargs):
     ALGORITHMS = {'numerical': 'polls.models.algorithm.NumericalComparisonAlgorithm',
-                  'string': 'polls.models.algorithm.StringComparisonAlgorithm', }
+                  'string': 'polls.models.algorithm.StringComparisonAlgorithm',
+                  'tree': 'polls.models.algorithm.DecisionTreeAlgorithm'
+                  }
     algorithm = class_import(ALGORITHMS[atype])(**kwargs)
     return algorithm
 
@@ -31,6 +35,7 @@ class Algorithm:
         raise NotImplementedError
 
 
+# not used
 class NumericalComparisonAlgorithm(Algorithm):
 
     name = 'numerical'
@@ -68,6 +73,7 @@ class NumericalComparisonAlgorithm(Algorithm):
         return json.loads(result.stdout)
 
 
+# not used
 class MutipleChioceComparisonAlgorithm(Algorithm):
     name = 'mutiple_choice'
     params = ('max_choice')
@@ -97,6 +103,7 @@ class MutipleChioceComparisonAlgorithm(Algorithm):
         return r
 
 
+# not used
 class MathExpressionComparisonAlgorithm(Algorithm):
     name = 'math_express'
     params = ('exclude')
@@ -183,8 +190,6 @@ class StringComparisonAlgorithm(Algorithm):
         return grade
 
 
-
-
 class AlgorithmField(models.Field):
     '''
     AlgorithmField will generate algorithm (Algorithm)by given
@@ -211,3 +216,160 @@ class AlgorithmField(models.Field):
         if isinstance(value, Algorithm):
             instance = algorithm_base_parser(instance)
         return json.dumps(instance)
+
+
+# Decision tree algorithm
+class DecisionTreeAlgorithm(Algorithm):
+
+    name = 'tree'
+    params = ('ignore_case', )
+
+    def __init__(self, **kwargs):
+        self.__args__ = {'ignore_case': False}
+        for k, v in kwargs.items():
+            if k in self.params:
+                self.__args__[k] = v
+
+    def deconstruct(self):
+        path = 'polls.models.algorithm.DecisionTreeAlgorithm'
+        args = [self.name]
+        kwargs = self.__args__
+        return (path, args, kwargs)
+
+    def run(self, tree, answer, args={}):
+        '''
+        answer: student answer,
+        tree: decision tree
+        return: result of processing tree
+        '''
+        return process_node(tree, answer, args)
+
+    def execute(self, result, args={}):
+        full = args["full"] if args["full"] else False
+        score = result["score"]
+        feedback = get_feedback(result, full)
+        return score, feedback
+
+
+class Node:
+    def __init__(self, node, NodeInput, args=None):
+        self.node = node
+        self.input = NodeInput
+        self.args = args
+
+    def decide(self, node):
+        assert node.get("type", 0) != 0  # don't decide a score node.
+
+        url = 'https://sagecell.sagemath.org'
+        code = node["title"]
+        code = code.replace("_value", self.input)  #
+        seed = self.args.get("seed", None)
+        script = self.args.get("script", "").replace("_value", self.input)
+        language = self.args.get("language", "sage")
+
+        if language == "maxima":
+            if seed:
+                pre = "_seed: {}$s1: make_random_state (_seed)$set_random_state (s1)$ ".format(seed)
+            code = "print(maxima.eval('{}'))".format(pre+script+" "+code)
+
+        else:
+            code = script+"\n"+code
+
+        pre = "_seed={}\nimport random\nrandom.seed(_seed)\n".format(seed)
+        sage = SageCell(url)
+        msg = sage.execute_request(pre+code)
+        try:
+            results = SageCell.get_results_from_message_json(msg)
+            if results == "True":
+                return True
+            elif results == "False":
+                return False
+            else:
+                raise ValueError('unexpected outcome from execution')
+        except ValueError as e:
+            raise e
+
+    def get_result(self):
+        if not self.node:  # handle some invalid cases
+            return 0, None
+        else:
+            self.node["state"] = 1  # visit node
+
+        if self.node["type"] == 0:  # we just need to return the score if it is a score node
+            return self.node
+        else:  # we need to process the decision first then go through its valid children.
+            # isRoot = False
+            children = self.node.get("children", [])
+            if self.node["type"] != -1:  # we don't decide root
+                myBool = self.decide(self.node)  # get condition
+                self.node["eval"] = myBool
+                bool_str = "true" if myBool else "false"
+                # decide feedback
+                feedback = self.node.get("feedback")
+                if feedback:
+                    feedback = feedback.get(bool_str)
+                    if feedback:
+                        self.node["feedback"] = feedback
+
+                # filter children
+                children = list(filter(lambda c: c['bool'] == myBool, children))
+                policy = self.node.get("policy")
+                policy = policy.get(bool_str, "sum") if policy else "sum"
+            else:
+                # isRoot = True
+                policy = self.node.get("policy", "sum")
+            # recursively get result from children, THIS CAN BE INPROVED BY BRANCH CUTTING
+            results = list(map(lambda c: process_node(c, self.input, self.args), children))
+            scores = list(map(lambda r: r["score"], results))
+
+            # based on the policy, get the score and return
+
+            if policy == "sum":
+                score = sum(scores)
+                self.node["score"] = score
+                for child in children:
+                    child["state"] = 2  # visited and used
+                return self.node
+
+            elif policy == "max":
+                score = max(scores)
+                self.node["score"] = score
+                index = scores.index(score)
+                children[index]["state"] = 2
+                return self.node
+
+            elif policy == "min":
+                print(scores)
+                score = min(scores)
+                self.node["score"] = score
+                index = scores.index(score)
+                children[index]["state"] = 2
+                return self.node
+
+
+def get_feedback(result, full=False):
+    state = result.get("state")
+    if not state:
+        return
+
+    feedbacks = []
+    current = result.get("feedback")
+    if current:
+        feedbacks.append(current)
+
+    if result["type"] == 0:
+        return feedbacks
+    else:
+        children = result.get("children")
+        if children:
+            for child in children:
+                if child.get("state") == 2 or full:
+                    feedback = get_feedback(child, full)
+                    if feedback:
+                        feedbacks += feedback
+        return feedbacks
+
+
+# We can use multiple threads to get the result
+def process_node(node, ProcInput, args):
+    return Node(node, ProcInput, args).get_result()
