+++ requirements.txt
@@ -5,8 +5,8 @@ dj-database-url==0.5.0
 Django>=2.2
 pylint==2.3.1
 autopep8==1.4.4
-psycopg2==2.7.7
-psycopg2-binary==2.7.7
+psycopg2==2.8.5
+psycopg2-binary==2.8.5
 django-cors-headers == 3.2.1
 Pillow >= 6.1.0
 django-ratelimit==2.0.0
